+++ native/features2d.h
@@ -7,6 +7,7 @@
 #include <opencv2/xfeatures2d.hpp>
 
 extern "C" {
+
 // =============================================================================
 //   MSER
 // =============================================================================
