+++ src/imgcodecs.rs
@@ -9,6 +9,18 @@ use std::os::raw::c_char;
 use std::path::Path;
 use failure::Error;
 
+extern "C" {
+    fn cv_imread(input: *const c_char, flags: ImageReadMode) -> *mut CMat;
+    fn cv_imdecode(buf: *const u8, l: usize, m: ImageReadMode) -> *mut CMat;
+    fn cv_imencode(
+        ext: *const c_char,
+        inner: *const CMat,
+        flag_ptr: *const ImageWriteMode,
+        flag_size: usize,
+    ) -> ImencodeResult;
+
+}
+
 // =============================================================================
 //  Imgproc
 // =============================================================================
@@ -106,18 +118,6 @@ pub enum ImageWritePngStrategy {
     /// Using this value prevents the use of dynamic Huffman codes, allowing for
     /// a simpler decoder for special applications.
     Fixed = 4,
-}
-
-extern "C" {
-    fn cv_imread(input: *const c_char, flags: ImageReadMode) -> *mut CMat;
-    fn cv_imdecode(buf: *const u8, l: usize, m: ImageReadMode) -> *mut CMat;
-    fn cv_imencode(
-        ext: *const c_char,
-        inner: *const CMat,
-        flag_ptr: *const ImageWriteMode,
-        flag_size: usize,
-    ) -> ImencodeResult;
-
 }
 
 #[repr(C)]
